finished the basic review_model, updated the freelancer_review and main_model models accordingly to go through this model or take out unnecessary code
